+++ CHANGES.md
@@ -4,3 +4,5 @@
 
 * User State Tracking
   * Now tracks the `properties_changed` event properly.
+* Shitlist
+  * Add avatar ID and assetbundle hash count to shitlist startup